refactor interfaces into packages
