--- fuseline/exporters.py
@@ -1,15 +1,0 @@
-from __future__ import annotations
-
-from .interfaces import Exporter
-from .workflow import Workflow, WorkflowSchema
-
-
-class YamlExporter(Exporter):
-    """Serialize workflows to a simple YAML format."""
-
-    def export(self, workflow: Workflow | WorkflowSchema, path: str) -> None:
-        """Serialize *workflow* or schema to YAML and write it to *path*."""
-
-        schema = workflow.to_schema() if isinstance(workflow, Workflow) else workflow
-        with open(path, "w", encoding="utf-8") as f:
-            f.write(schema.to_yaml())
